+++ pkg/project/project.go
@@ -3,7 +3,7 @@ package project
 var (
 	buildTimestamp string
 	gitSHA         string
-	version        = "8.2.1-dev"
+	version        = "8.2.2-dev"
 )
 
 func BuildTimestamp() string {